pareint

+++ index.js
@@ -149,7 +149,7 @@ const getTeamReport = async (text_date, event) => {
         const querySnapshot = await db.collection('counting').where('team', '==', runner_db.team).where('date', '==', callDate(text_date)).get();
         querySnapshot.forEach((doc) => {
             console.log(doc.id, ' => ', doc.data())
-            let step = parseFloat(doc.data().step)
+            let step = parseInt(doc.data().step)
             sum_step += step
             text_reply += `${doc.data().name} เดินไป ${doc.data().step} ก้าว\n`
             round += 1;
